requirements.txt
discord
python-dotenv
jsonpickle